Rework Code + Deactivated Mods can now be changed too

+++ QModManager/Utility/IOUtilities.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using QModManager.DataStructures;
+using QModManager.Patching;
 using Oculus.Newtonsoft.Json;
 
 namespace QModManager.Utility
@@ -128,36 +128,45 @@ namespace QModManager.Utility
             return s;
         }
 
-        internal static void ChangeModStatustoFile(SimpleModDataTemplate smdt)
+        internal static void ChangeModStatustoFile(QMod qmod)
         {
-            //Get the Configfile
-            string modconfigpath = Path.Combine(Path.GetFullPath(Path.GetDirectoryName(smdt.PathToAssemblyFile)), "mod.json");
-            var modconfig = JsonConvert.DeserializeObject<Dictionary<string, object>>(File.ReadAllText(modconfigpath));
-
-            //Modify the Configfile
-            foreach (var kvp in modconfig)
+            //Get the Configfile (The mod.json in the Mod Directory)
+            string modconfigpath = Path.Combine(Path.GetFullPath(qmod.SubDirectory), "mod.json");
+            if (File.Exists(modconfigpath))
             {
-                if (kvp.Key.ToLower() == "enable")
+                //we doing it with a Dictionary instead of <IQMod> because we have Data in the File that is NOT handles by QMM we would loose this information and that is bad.
+                var modconfig = JsonConvert.DeserializeObject<Dictionary<string, object>>(File.ReadAllText(modconfigpath));
+
+                //Modify the Configfile
+                foreach (var kvp in modconfig)
                 {
-                    modconfig[kvp.Key] = smdt.Enabled;
-                    break;
+                    if (kvp.Key.ToLower() == "enable")
+                    {
+                        modconfig[kvp.Key] = qmod.Enable;
+                        break;
+                    }
+                }
+
+                //Create a JSON Format so it will be Readable by User in a Text Editor
+                Formatting myformat = new Formatting();
+                myformat = Formatting.Indented;
+
+                //Save it back to File
+                string jsonstr = JsonConvert.SerializeObject(modconfig, myformat);
+                try
+                {
+                    File.WriteAllText(modconfigpath, jsonstr);
+                    Logger.Log(Logger.Level.Info, $"IOUtilities - ChangeModStatustoFile - Enabled Status Update for {qmod.Id} was succesful written to: {modconfigpath}");
+                }
+                catch
+                {
+                    Logger.Log(Logger.Level.Error, $"ErrorID:5713/36B - Saving mod.json for Mod {qmod.Id} to {modconfigpath} failed. - Was the File open in a other Program ? Permission Error ?");
                 }
             }
-
-            //Save it back
-            Formatting myformat = new Formatting();
-            myformat = Formatting.Indented;
-            string jsonstr = JsonConvert.SerializeObject(modconfig, myformat);
-            try
+            else
             {
-                File.WriteAllText(modconfigpath, jsonstr);
-                Logger.Log(Logger.Level.Info, $"IOUtilities - ChangeModStatustoFile - mod.json Update for {smdt.ID} was succesful saved to Mod Folder");
+                Logger.Log(Logger.Level.Error, $"ErrorID:5713/17A - File {modconfigpath} does not exist! - Mod {qmod.Id}");
             }
-            catch
-            {
-                Logger.Log(Logger.Level.Error, "ErrorID:5713/31A - Saving Changed Mod Configfile failed");
-            }
-
         }
     }
 }
